refactor: Modularize sidebar and checkbox logic
Refactored the sidebar and checkbox functionality to improve modularity and maintainability.

- Encapsulated the sidebar toggle logic into a new `src/modules/sidebar.js` module, removing it from the main `app.js` file.
- Replaced the hardcoded, ID-based mutual exclusivity for checkboxes with a scalable, data-driven approach using `data-exclusive-group` attributes. This simplifies adding new exclusive checkbox groups, as no JavaScript changes are required.
